RL: RLDriver should use mvconfdict to create the multi-view NN.
At the same time, the experiments have shown conf.SV_VISCFG performs
reasonably well for multi-view NN (by mistake). So we keep this
configuration in cat.py and we can evaluate the classifier without
retraining the whole net.

For RL we increase the number of channels for (potentially)
better prediction.