RL: add --curiosity_type to support collision based curiosity (draft).
Motivation: the feature based curiosity usually gives large rewards, and
since there is no upper limit on the domain of feature vector.
Experiments also showed large L2 norm (around 4000), and failed to
converge to small (< 1.0) values during the whole training process.
We had to use large rewards (1e3 to 1e7) as well to compensate the negative impacts
from this curiosity-based artificial reward.

Solution: instead of comparing feature vectors directly, we modified the
forward model to predict the ratio of the trajectory performed according
to the feature vector and the action distribution. Since the ratio is
within [0.0,1.0], we used sigmoid as the last layer.