build: introduce MKOBS_USE_CGAL so user can disable items depending on CGAL.
This is crucial for apps/libs that use Eigen 3.3, CGAL 4.7 with and libigl,
because libigl's meshbool is not compatitable with the above combination.

Working case:
1. Eigen 3.3 and CGAL 4.11 (Fedora 28)
2. Eigen 3.2 and CGAL 4.7 (Ubuntu 16.04)

Currently we are heavily relying on Eigen 3.3's new features so downgrading is
not a cheap option.

Note: Ubuntu enables GMP (GNU Multiple Precision arithmetic library) by
default. This might also be the of troubles.