RL/curiosity-rl: implement --train tunnel_finder_foreach1
This method creates a network to predict the delta states from the input image
to all known narrow tunnel vertices. The intuition is to simplify the
model for each individual delta state prediction

We may run an experiment about how this works for predicting one narrow tunnel
vertex, because multiple predictions share the most of the network and hence
overfitting is hard.