lib/osr,pyosr: implement pyosr.UnitWorld.enum_2drot_free_configuration
This function enumerate 2D (altitude and azimuth) rotations while keep
the robot point aligning with the environment point + margin *
environment surface normal.

The margin is calculated to ensure the two triangles where the robot and
environment points live on will not collide with
each other.

However it is quite common that the margin is too small for valid
configurations. We need to seek for better alternatives.